Entirely replaced cur.c
The new cur.c is significantly nicer. After ui, this was the
second worst bit of code. Next will be to replace buf.c with a
buf/text.c, buf/line.c, and buf/buf.c files. This will all be
built in a thread safe way to allow for multiple threads to do
things, including, importantly, a proper rendering thread(!!),
which would probably almost double performance already.

This is the first part of a big restructure. Also note that
regions are broken rn.