Added visible cursor for bar mode.
This got removed when I changed over to a seprate bar.c
The new cursor is blue at the end of the line and black
otherwise. I decided to put this cursor back when a few times
I lost tack of my mode and didn't realise I was in bar mode.