Added inp_key translation table
Before keys are recieved, they now go through a translation
table. This allows non-escape-code keys to turn into special
key codes, in the 0x1.. form. This makes special keys a little
nicer, and since the key translations are kept in a table, it
is more flexible and can probably be modified at runtime soon.